Création des traits
