+++ src/Entity/Breeder.php
@@ -13,9 +13,6 @@ class Breeder extends User
     #[ORM\Column]
     private ?bool $isAdmin = null;
 
-    #[ORM\Column(length: 255)]
-    private ?string $name = null;
-
     #[ORM\OneToMany(mappedBy: 'breeder', targetEntity: Offer::class, orphanRemoval: true)]
     private Collection $offers;
 
@@ -33,18 +30,6 @@ class Breeder extends User
     public function setIsAdmin(bool $isAdmin): self
     {
         $this->isAdmin = $isAdmin;
-
-        return $this;
-    }
-
-    public function getName(): ?string
-    {
-        return $this->name;
-    }
-
-    public function setName(string $name): self
-    {
-        $this->name = $name;
 
         return $this;
     }
