+++ src/Entity/Traits/HasDescrTrait.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace App\Entity\Traits;
+
+use Doctrine\ORM\Mapping as ORM;
+
+trait HasDescrTrait
+{
+    #[ORM\Column(type: Types::TEXT)]
+    protected ?string $description = null;
+
+    public function getDescription(): ?string
+    {
+        return $this->description;
+    }
+
+    public function setDescription(string $description): self
+    {
+        $this->description = $description;
+
+        return $this;
+    }
+}
